# dakunlun/docs/swagger.yaml
definitions:
...
  entity.Skill:
    properties:
      id:
        type: integer
      level:
        type: integer
    type: object
  entity.Skills:
    additionalProperties:
      $ref: '#/definitions/entity.Skill'
    type: object
  entity.Skin:
    properties:
      active:
        type: boolean
      fightingCapacity:
        type: integer
      val:
        type: integer
    type: object
  entity.SkinMap:
    additionalProperties:
      $ref: '#/definitions/entity.Skin'
    type: object
...
  msg.Hero:
    properties:
      attackFreq:
        type: integer
      attackTrans:
        type: integer
      defendTrans:
        type: integer
      evolveTimes:
        type: integer
      exploreID:
        type: integer
      fightingCapacity:
        type: integer
      heroID:
        type: integer
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      race:
        type: integer
      sex:
        type: integer
      skills:
        $ref: '#/definitions/entity.Skills'
      skinID:
        type: integer
      skins:
        $ref: '#/definitions/entity.SkinMap'
      type:
        type: integer
    type: object
...
  msg.HeroEvolveResponse:
    properties:
      evolveTimes:
        type: integer
    type: object
  msg.HeroJoinResponse:
    type: object
  msg.HeroListResponse:
    properties:
      heros:
        items:
          $ref: '#/definitions/msg.Hero'
        type: array
    type: object
  msg.HeroSkinGetResponse:
    properties:
      skin:
        $ref: '#/definitions/entity.Skin'
    type: object
  msg.HeroSkinUseResponse:
    properties:
      fightingCapacity:
        type: integer
      skinID:
        type: integer
    type: object
  msg.HeroUnloadResponse:
    type: object
  msg.HeroUpgradeResponse:
    properties:
      level:
        type: integer
    type: object
  msg.LoginResponse:
    properties:
      refreshToken:
        description: REFRESHTOKEN（TOKEN快过期时延期用）
        type: string
      refreshTokenExpireTime:
        description: REFRESHTOKEN过期时间
        type: integer
      token:
        description: TOKEN
        type: string
      tokenExpireTime:
        description: TOKEN过期时间
        type: integer
      uid:
        description: 用户ID
        type: integer
    type: object
...
host: 8.140.113.88:8080
info:
  contact: {}
  description: 接口文档
  title: dakunlun
  version: "1.0"
paths:
...
  /api/hero/evolve:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 伙伴id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.HeroEvolveResponse'
      summary: 伙伴锻魂
  /api/hero/join:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 伙伴id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.HeroJoinResponse'
      summary: 设置伙伴出战
  /api/hero/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.HeroListResponse'
      summary: 伙伴列表
  /api/hero/skin/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 伙伴id
        in: query
        name: id
        required: true
        type: integer
      - description: 皮肤ID
        in: query
        name: skinID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.HeroSkinGetResponse'
      summary: 看皮肤广告
  /api/hero/skin/use:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 伙伴id
        in: query
        name: id
        required: true
        type: integer
      - description: 皮肤IO
        in: query
        name: skinID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.HeroSkinUseResponse'
      summary: 使用皮肤
  /api/hero/unload:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.HeroUnloadResponse'
      summary: 设置伙伴下阵
  /api/hero/upgrade:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 伙伴id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.HeroUpgradeResponse'
      summary: 伙伴升级
...
  /login/login/native:
    post:
      consumes:
      - application/json
      parameters:
      - description: 设备ID
        in: query
        name: name
        required: true
        type: string
      - description: 是否清除账号
        in: query
        name: clear
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.LoginResponse'
      summary: 登录
...
swagger: "2.0"